+++ Source/Plugins/Sampling/SamplingFilters/CropSurfaceMesh.cpp
@@ -320,7 +320,7 @@ void CropSurfaceMesh::execute()
   Int64ArrayType::Pointer newNodeNumbers = Int64ArrayType::CreateArray(numVerts, "newNumbers");
   newNodeNumbers->initializeWithValue(-1);
   int64_t* newNNumbers = newNodeNumbers->getPointer(0);
-  for (size_t i = 0; i < numVerts; i++)
+  for (int64_t i = 0; i < numVerts; i++)
   {
     badNode = false;
     if(vertex[i*3] < m_XMin || vertex[i*3] > m_XMax) { badNode = true; }
@@ -355,7 +355,7 @@ void CropSurfaceMesh::execute()
   Int64ArrayType::Pointer newFaceNumbers = Int64ArrayType::CreateArray(numFaces, "newNumbers");
   newFaceNumbers->initializeWithValue(-1);
   int64_t* newFNumbers = newFaceNumbers->getPointer(0);
-  for (size_t i = 0; i < numFaces; i++)
+  for (int64_t i = 0; i < numFaces; i++)
   {
     badFace = false;
     node1 = face[i*3];
@@ -409,7 +409,7 @@ void CropSurfaceMesh::execute()
     }
 
     // Find the unique set of feature ids
-    for (size_t i = 0; i < numFaces; ++i)
+    for (int64_t i = 0; i < numFaces; ++i)
     {
       int32_t currentFeatureId = m_FeatureIds[i];
       if (currentFeatureId < totalFeatures)
